pyproject.toml
[tool.poetry]
name = "image-content-moderation"
version = "0.1.0"
description = "Image content moderation service — detects symbols, flags, banners, and text markers on images"
authors = ["Dev Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
python-multipart = "^0.0.18"
pydantic = "^2.9.0"
pydantic-settings = "^2.5.0"
structlog = "^24.4.0"
Pillow = "^10.4.0"
torch = "^2.4.0"
torchvision = "^0.19.0"
easyocr = "^1.7.0"
ftfy = "^6.2.0"
regex = "^2024.7.0"
clip = {git = "https://github.com/openai/CLIP.git"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.0"
ruff = "^0.6.0"

[tool.poetry.scripts]
start = "uvicorn app.main:create_app --factory --host 0.0.0.0 --port 8000"

[tool.ruff]
target-version = "py39"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "ANN", "B", "SIM"]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

